DRY up reporter-trimming code by placing it in a mixin class
